+++ Automation/CHANGELOG.md
@@ -1,5 +1,6 @@
 # v4.9.0 (TBD)
 * [Feature] Allow combining conditions with AND and OR in the rule constructor.
+* [Feature] Add an option to show optional parentheses around nested logical groups in rule descriptions.
 
 # v4.8.1 (July 18th, 2026)
 * [Update] Support Timberborn `1.1.1.1`.